completed unit test

+++ backend/tasks/views/delete_task.py
@@ -6,7 +6,7 @@
 """
 
 from django.contrib.auth.mixins import LoginRequiredMixin
-from django.http import HttpResponseRedirect, Http404
+from django.http import HttpResponseRedirect, Http404, HttpResponseForbidden
 from django.urls import reverse
 from django.views import View
 
@@ -23,7 +23,7 @@ class TaskDeleteView(LoginRequiredMixin, View):
             task = Task.objects.get(id=task_id)
         except Task.DoesNotExist:
             raise Http404()
-        except Task.MultipleObjectsReturned:
-            raise Http404()
+        if task.author != request.user:
+            return HttpResponseForbidden()
         task.delete()
         return HttpResponseRedirect(reverse('tasks:tasklist'))
